requirements.txt
flask
pandas
scikit-learn
flask-cors
python-dotenv
pyjwt
bcrypt
